Moved all SERVER files to proper place for manual import
