--- imports/api/filmManagement/comments/server/comments.js
@@ -1,41 +1,0 @@
-Meteor.publish( 'post', function( postId ) {
-  check( postId, String );
-
-  if ( postId ) {
-    return [
-      Users.find( { '_id': postId } ),
-      Comments.find( { 'postId': postId }, { sort: { "date": -1 } } )
-    ];
-  } else {
-    return null;
-  }
-});
-
-
-// Meteor.publish( 'post', function( postId ) {
-//   check( postId, String );
-//
-//   if ( postId ) {
-//     return Users.find( { '_id': postId } );
-//   } else {
-//     return null;
-//   }
-// });
-
-// Meteor.publish( 'comments', function( postId ) {
-//
-//   check( postId, String );
-//
-//   if ( postId ) {
-//     return Comments.find( {}, { sort: { "date": -1 } } );
-//   } else {
-//     return null;
-//   }
-// });
-
-
-// Meteor.publish( 'comments', function(postId) {
-//      return Comments.find({}, {sort: {date: -1}});
-//      console.log("EXEC!");
-// });
-//
