+++ imports/api/filmManagement/privateMessages/methods.js
@@ -1,0 +1,3 @@
+import './privateMessages.js';
+
+// privateMessages has no Meteor.methods
