+++ imports/api/filmManagement/privateMessages/privateMessages.js
@@ -1,7 +1,9 @@
+import { Mongo } from 'meteor/mongo';
+
 PrivateMessages = new Mongo.Collection('privateMessages');
 
-if (Meteor.isServer) {
-    Meteor.publish('privateMessages.all',function(){
-        return PrivateMessages.find({});
-    });
-}
+PrivateMessages.allow({
+    insert: function(userId,doc) { return true; },
+    update: function(userId,doc,fieldNames, modifier) { return true; },
+    remove: function(userId,doc) { return false; }
+});
