--- imports/api/filmManagement/privateMessages/server/privateMessage.js
@@ -1,13 +1,0 @@
-Meteor.startup(function(){
-   PrivateMessages.allow({
-        insert:function(userId,doc){
-            return true;
-        },
-        update:function(userId,doc,fieldNames, modifier){
-            return true;
-        },
-        remove:function(userId,doc){
-            return false;
-        }
-    }); 
-});
