+++ imports/api/filmManagement/users/server/create-user.js
@@ -1,12 +1,7 @@
 Accounts.onCreateUser(function(options, user) {
     if (options.profile && options.profile.user) {
-
         var optionsUser = options.profile.user;
-
-
-
         optionsUser.userId = user._id;
-
         user.profile = options.profile;
     }
 
@@ -15,6 +10,4 @@ Accounts.onCreateUser(function(options, user) {
     }
 
     return user;
-
-
 });
